Fix bug where bounds() would be invalid if diff had only geometry changes
The bug was introduced in 8d512bb. The intent of that commit was return
a more sensible bounding box for changesets where edits in a small area
happened to alter the geometry of a very large way. The naive behavior
(returning a bounding box that encompasses all modified elements) would
then be zoomed very far out, obscuring the most interesting changes.

This commit introduces a more sophisticated (perhaps overcomplicated)
approach, where we compute both the bounding box of all changed features
as well as the bounding box of the subset that have tag changes. We return
the former bounding box _unless_ the latter bounding box is well-defined
and also much smaller, in which case we return it instead.